+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - New action: "Go to {systemName}": Go to the URL defined by `siteUrl`
+- Show the badge count when rendering utility items (for actions like "Updates")
 
 ## 1.0.0 - 2022-02-03